requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0